Project tree restructuration
