+++ helpers/create_rental.rb
