+++ helpers/create_teacher.rb
