+++ helpers/nameable.rb
