+++ helpers/rental.rb
