+++ main.rb
@@ -1,8 +1,5 @@
 #!/usr/bin/env ruby
 require 'pry'
-require './teacher'
-require './student'
-require './book'
 require './app'
 
 def main
